minimal oprydning af bindmixer og innkeeper delt i 5

+++ tavern.h
@@ -6,9 +6,9 @@ private:
     std::string localFood0 = "Pork";
     std::string localFood1 = "Venison";
     std::string localFood2 = "Lamb";
-    int small;
-    int medium;
-    int large;
+    double small;
+    double medium;
+    double large;
     std::string mead;
 public:
     std::string getFood(int menuItem);
